Fixes
Made some changes in the layout of how the entries are seen before clicked on. Changed some CSS. May get rid of robotics page altogether and put Marvin in Uni projects as the Marvin project may be the only one ever put onto this page and don't want wasted space.